azure-akscore-e2e-test/data-loader/requirements.txt
azure-storage-blob==12.20.0
requests==2.32.3
